feat: add online\offline status to manager and datastore (#29)
Added a new backend api to perform a simple query of the redfish root service for a blade\host device.  This allows for a quick check to see if the device, after previously being successfully added to the service, is still present.
This quick check translates directly to an online\offline status for the device.  This is mainly used to handle the situation when a blade\host get powered down.
This status is then used to modify the service's behavior for that device.  The status state is also returned via the client so that the webui\cli can display useful information to the user about the device's current state.