fix(service): fix error handling bugs during add blade\host
Found a couple bugs in appliance.AddBlade() and manager.AddHost().
(1) corner case where an empty bladeId is being allowed, when it shouldn't
(2) Not enough information is being saved to new blade objects during error handling.  This was causing a panic during the delete because the backendOps variable was a nil point being used to call a backend function.